Bug 265772 - [wsdl4j] Upgrade to WSDL4J 1.6
This change allows all of webtools to standardise on a single
version of wsdl4j (1.6.2). The necessary unimplmented methods
are implemented and the bundle manifest requirements are
updated accordingly.

We no longer need the wsdl15 feature, which seems to exist solely
so it can provide WSDL4J 1.5.1, so this change also disables the
build of that feature and it is no longer included in the update
site.

Change-Id: Ibacfd65c324a7bf635387388f9bffaa4efe7fc4d
Signed-off-by: Mat Booth <mat.booth@redhat.com>